implement run-time specialisation of vector_multiply_add functions
Compile relevant source files multiple times with different SIMD
extensions enabled and with a different preprocessor define which sets
the Feature enum in template specialisations.

This feels a bit messy, but it means we can use existing code which uses
__AVX__ type flags to detect the implementation at compile time, but
select between them at run-time.

In fact if we want to use intrinsics this is the only way (i.e. one
compilation unit per specialisation), because if a given extension is
enabled then the compiler may use instructions from that extension in
the whole compilation unit, so it might not be safe to call functions in
it even if they only use intrinsics from extensions that are available
at run-time.

The rationale for the set of feature flags used is:

- AVX2 isn't really useful for us, but MSVC bundles together FMA and
  AVX2, and only defines __AVX2__, not __FMA__. In practice, these are
  bundled together in CPUs too.

  When using FMA we detect __AVX2__, but also check at run-time if we
  have both FMA and AVX2.

- sse4.2 has been around since nehalem, 2008

- MSVC with no flags supports SSE4.2

only tested on linux so far